Pull out CardSelector as generic component
